introduce mockito in bookkeeper server
